before trying tabele editing

+++ lib/model/purchase_order_table_model.dart
@@ -11,7 +11,7 @@ class PurchaseOrderTableModel with _$PurchaseOrderTableModel {
     //final String? vat,
     @JsonKey(name: "purchase_uom") final String? purchaseUom,
     @JsonKey(name: "purchase_uom_name") final String? purchaseUomName,
-   @JsonKey(name: "excess_tax") final int? excessTax,
+   @JsonKey(name: "excess_tax") final double? excessTax,
     @JsonKey(name: "unit_cost") final double? unitCost,
    // @JsonKey(name: "return_type") final String? returnType,
    // @JsonKey(name: "return_time") final String? returnTime,
